+++ src/App.jsx
@@ -1,23 +1,15 @@
 import React from 'react';
-
-const todoList = [
-  { id: 1, title: 'Complete assignment' },
-  { id: 2, title: 'Buy groceries' },
-  { id: 3, title: 'Attend meeting' }
-];
+import TodoList from './TodoList';
+import AddTodoForm from './AddTodoForm';
 
 function App() {
   return (
     <div>
       <h1>Todo List</h1>
-      <ul>
-        {todoList.map((todo) => (
-          <li key={todo.id}>{todo.title}</li>
-        ))}
-      </ul>
+      <AddTodoForm />
+      <TodoList />
     </div>
   );
 }
 
 export default App;
-
